Hiển thị danh sách giỏ hàng, chỉnh sửa số lượng, xóa 1 sản phẩm, xóa tất cả sản phẩm

+++ src/items/CartItem/ItemCart.js
@@ -1,14 +1,13 @@
-import {View, Text, TouchableOpacity, Image, ToastAndroid} from 'react-native';
-import React, {useEffect, useRef, useState} from 'react';
-import {itCart} from './style';
+import { View, Text, TouchableOpacity, Image, ToastAndroid } from 'react-native';
+import React, { useEffect, useRef, useState } from 'react';
+import { itCart } from './style';
 import appst from '../../constants/AppStyle';
 import Swipeable from 'react-native-gesture-handler/Swipeable';
 import SweetAlert from 'react-native-sweet-alert';
-import {deleteOneItemCard, updateCartItem} from '../../api/CartApi';
+import { deleteOneItemCard, updateCartItem } from '../../api/CartApi';
 
 const ItemCart = ({
   item,
-  cards,
   setCards,
   setTotalPrice,
   currentlyOpenSwipeable,
@@ -17,6 +16,7 @@ const ItemCart = ({
   setCheckedProducts,
   isChecked,
   onCheckedChange,
+  refreshCart,
 }) => {
   const [productQuantity, setProductQuantity] = useState(item.quantity);
 
@@ -39,7 +39,7 @@ const ItemCart = ({
 
   const calculateTotalPrice = products => {
     const total = products.reduce((sum, product) => {
-      return sum + product.product_id.price * product.quantity;
+      return sum + product.price * product.quantity;
     }, 0);
     setTotalPrice(total);
   };
@@ -72,9 +72,10 @@ const ItemCart = ({
         },
         async () => {
           setProductQuantity(0);
-          await updateQuantity(0); // Remove item
+          await updateQuantity(0);
           setCheckedProducts(prev => prev.filter(cart => cart._id !== item._id));
           setCards(prev => prev.filter(cart => cart._id !== item._id));
+          refreshCart();
         },
       );
     }
@@ -91,6 +92,7 @@ const ItemCart = ({
       const response = await updateCartItem(productData);
       if (response.status) {
         console.log('Cập nhật số lượng giỏ hàng thành công');
+        refreshCart();
       } else {
         console.log('Cập nhật số lượng giỏ hàng thất bại');
       }
@@ -108,14 +110,14 @@ const ItemCart = ({
   };
 
   const imageAssets =
-  item &&
-  item.product_id &&
-  item.assets.filter(asset => {
-    return asset.match(/\.(jpeg|jpg|png|gif)$/);
-  });
+    item &&
+    item.product_id &&
+    item.assets.filter(asset => {
+      return asset.match(/\.(jpeg|jpg|png|gif)$/);
+    });
 
-const imageUrl =
-  imageAssets && imageAssets.length > 0 ? imageAssets[0] : 'https://i.pinimg.com/236x/6a/f1/ec/6af1ec6645410a41d5339508a83b86f9.jpg'; // Giá trị mặc định nếu không có ảnh
+  const imageUrl =
+    imageAssets && imageAssets.length > 0 ? imageAssets[0] : 'https://i.pinimg.com/236x/6a/f1/ec/6af1ec6645410a41d5339508a83b86f9.jpg'; // Giá trị mặc định nếu không có ảnh
 
   const swipeableRef = useRef(null);
 
@@ -125,22 +127,29 @@ const imageUrl =
         product_id: item.product_id,
         size_name: item.sizeName,
       };
-      console.log('body delete cart: ', body);
-
+  
+      console.log('Body delete cart: ', body);
+      
+      if (!body.product_id || !body.size_name) {
+        console.log('Missing product_id or size_name:', body);
+        return;
+      }
+  
       const response = await deleteOneItemCard(body);
       if (response.status) {
         setCards(prevCards => prevCards.filter(cart => cart._id !== item._id));
         setCheckedProducts(prevChecked =>
           prevChecked.filter(cart => cart._id !== item._id),
         );
-        ToastAndroid.show('Da xoa sp', ToastAndroid.SHORT);
+        ToastAndroid.show('Đã xóa sản phẩm', ToastAndroid.SHORT);
+        refreshCart()
       } else {
-        ToastAndroid.show('Loi server', ToastAndroid.SHORT);
+        ToastAndroid.show('Lỗi server', ToastAndroid.SHORT);
       }
     } catch (error) {
-      console.log('error delete item card->', error);
+      console.log('Error delete item card: ', error);
     }
-  };
+  };  
 
   const rightSwipeable = () => {
     return (
